changed property type

+++ OnlinePortalBackend/DTOs/ReceivingDTOs/ServiceRatingReceivingDTO.cs
@@ -7,7 +7,7 @@ namespace OnlinePortalBackend.DTOs.ReceivingDTOs
     {
         public long CustomerDivisionId { get; set; }
         public long ServiceId { get; set; }
-        public long Rating { get; set; }
+        public decimal Rating { get; set; }
         public int Recommendation { get; set; }
         public string Review { get; set; }
     }
@@ -16,7 +16,7 @@ namespace OnlinePortalBackend.DTOs.ReceivingDTOs
     {
         public long CustomerDivisionId { get; set; }
         public long ApplicationId { get; set; }
-        public long Rating { get; set; }
+        public decimal Rating { get; set; }
         public string Review { get; set; }
     }
